requirements.docker.txt
Django == 2.1.7

djangorestframework == 3.9.2
markdown == 3.1

uWSGI == 2.0.18
